SUS-662 escape values of HTML attributes
The test failed while comparing a string with its representation as a value of an HTML attribute.

The expected value was `"Download" Adult Coloring Book Stress Relieving Cats PDF Free`.

The actual value was `&quot;Download&quot; Adult Coloring Book Stress Relieving Cats PDF Free`.

It is a defect of the test not the code being tested. The expected value should be first escaped.