FAN-1285 Fix broken tests after updating global search links
On corporate pages, when showing English nav, the search now points to the Stories and Communities search.

This set of changes updates the tests so they can understand the first results as shown on that page. For the tests testing multiple results the tests can now navigate to the "More communities" search page which is basically the old page.