(PLATFORM-3465) Fix unpin button selector for non-English wikis (#2313)
Anons will see a wiki's i18n messages in the content language of the
wiki, so selecting based on the content of an attribute that uses an
i18n message won't work on a non-English wiki such as those in our
language path tests. Additionally, selecting based on a translation
string is brittle, so let's use a class selector here.